Emit array diff in order
Fixes #6

When emitting a patch file for an array, the elements are emitted in
reverse order.  This causes the application of the patch file to fail
because it tries to add elements beyond the end of the array (initially
empty).

This change emits the diff elements of an array in order.