Remove web/serve/files.go from source control.
The web UI is included in main.go but the WASM files must be generated
and packed first. `go get` (reasonably) doesn't run arbitrary commands
so the binary must also compile without the web UI files.

This change adds a build tag to web/serve/serve.go (and the generated
files.go) so it must be explicitly included with `-tags
include_web`. A single file plugin.go contains the Handle binding that
main.go depends on. If serve.go is included in the build it will bind
the actual handler to the exported variable. main.go just checks if
the handler is nil or not before starting the web uI.

This allows the generated web UI files to be removed from source
control and built on demand with the Makefile. `make release` will
include the web UI.